requirements.txt
distlib==0.3.8
dnspython==2.6.1
filelock==3.13.1
platformdirs==4.1.0
pymongo==4.6.2
setuptools==69.0.3
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
wheel==0.42.0
fastapi
uvicorn
matplotlib==3.8.2
numpy==1.26.4
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
python-multipart